Extract data query tests into separate test file.
